Properly sanitize murano package-list output before using it
In some cases, the murano package-list output may include a warning [1],
and that confuses the provider. We need to sanitize the output before
using it.

This time, we are using the right syntax for the new method.

[1] - https://review.rdoproject.org/etherpad/p/murano-failure, seen
      with urllib3 1.21.1

Change-Id: I01f750908540e5036e46a31bfaf49e655c8f37ca
(cherry picked from commit 6ab2fdc82605906f2acf95931e2fc8c7d388764b)